+++ src/components/Ballance/Ballance.jsx
@@ -1,3 +1,5 @@
+import css from "./Ballance.module.css"
+
 const Ballance = () => {
     return <div>Ballance-блок відображення балансу юзера</div>;
 };
