+++ src/components/Currency/Currency.jsx
@@ -1,3 +1,5 @@
+import css from "./Currency.module.css"
+
 const Currency = () => {
     return <div>Currency - блок відображення валюти</div>;
 };
