+++ src/components/Navigation/Navigation.jsx
@@ -1,3 +1,5 @@
+import css from "./Navigation.module.css"
+
 const Navigation = () => {
     return <div>Navigation - блок навігагації</div>;
 };
