--- src/components/Sidebar.jsx
@@ -1,13 +1,0 @@
-const Sidebar = () => {
-    return <div>
-        <p>Sidebar</p>
-        <Navigation />
-        <div>
-            <Ballance />
-            <Currency />
-        </div>
-    </div>;
-};
-export default Sidebar;
-
-
